uxmhf: sentinel,uhmpgtbl: obtain page-table base via uhmpgtbl_getmpgtblbase uapi within sentinel and dump contents along with regular page-table base

+++ uxmhf/xmhf-uobjs/include/uapi_uhmpgtbl.h
@@ -79,7 +79,7 @@ typedef struct {
 
 typedef struct {
     u32 dst_slabid;
-    u64 mpgtblbase;
+    u32 mpgtblbase;
 }__attribute__((packed)) uapi_uhmpgtbl_getmpgtblbase_params_t;
